SE MEJORO EL MANEJO DE DATOS PARA GENERAR EL PDF

+++ src/main/java/com/actinver/report_generator/dto/CarteraDetalleDTO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class CarteraDetalleDTO {
 	
 	private String descripcion;
-	private Long porcentaje;
+	private Double porcentaje;
 
 }
